Move output to separate module
